Run some checks every 4 hours.
