+++ .github/workflows/status.yml
@@ -1,10 +1,10 @@
-name: CI
+name: Status
 on:
-  - pull_request
-  - push
+  schedule:
+    - cron:  '13 */4 * * *'
 jobs:
   tests:
-    name: Add-on testing
+    name: Checks
     runs-on: ubuntu-latest
     env:
       ADDON_USERNAME: ${{ secrets.ADDON_USERNAME }}
@@ -17,31 +17,16 @@ jobs:
       KODI_STUB_RPC_RESPONSES: ${{ github.workspace }}/tests/rpc
     strategy:
       fail-fast: false
-      matrix:
-        python-version: [2.7, 3.5, 3.6, 3.7, 3.8]
     steps:
       - name: Check out ${{ github.sha }} from repository ${{ github.repository }}
         uses: actions/checkout@v2
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v1
-        with:
-          python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          sudo apt-get install gettext
-          sudo pip install coverage --install-option="--install-scripts=/usr/bin"
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-      - name: Run pylint
-        run: make check-pylint
-      - name: Run tox
-        run: make check-tox
-      - name: Check translations
-        run: make check-translations
-      - name: Run unit tests
-        run: coverage run -m unittest discover
-      - name: Run addon service
-        run: timeout --preserve-status -s SIGINT 10 coverage run -a service.py
-      - name: Upload code coverage to CodeCov
-        uses: codecov/codecov-action@v1
-        continue-on-error: true
+      - name: TEST Authentication
+        run: python -m unittest -v test_auth
+      - name: TEST TV guide
+        run: python -m unittest -v test_epg.TestEpg.test_get_epg
