+++ tests/test_epg.py
@@ -17,11 +17,11 @@ kodi = KodiWrapper()
 routing = plugin.routing
 
 
-class TestVtmGoEpg(unittest.TestCase):
+class TestEpg(unittest.TestCase):
     """ Tests for VTM GO EPG API """
 
     def __init__(self, *args, **kwargs):
-        super(TestVtmGoEpg, self).__init__(*args, **kwargs)
+        super(TestEpg, self).__init__(*args, **kwargs)
 
         self._vtmgoepg = vtmgoepg.VtmGoEpg(kodi)
